AS7-5375 CLI: unify the format of -l output between the deploy and undeploy commands
